+++ src/buffer.c
@@ -126,7 +126,7 @@ Buffer *find_buffer(const char *abs_filename)
     struct stat st;
     bool st_ok = stat(abs_filename, &st) == 0;
 
-    for (int i = 0; i < buffers.count; i++) {
+    for (size_t i = 0; i < buffers.count; i++) {
         Buffer *b = buffers.ptrs[i];
         const char *f = b->abs_filename;
 
@@ -142,7 +142,7 @@ Buffer *find_buffer(const char *abs_filename)
 
 Buffer *find_buffer_by_id(unsigned int id)
 {
-    for (int i = 0; i < buffers.count; i++) {
+    for (size_t i = 0; i < buffers.count; i++) {
         Buffer *b = buffers.ptrs[i];
         if (b->id == id) {
             return b;
